requirements.txt
aiohttp[speedups]
aiogram
opencv-python
peewee
Pillow
pytesseract
psycopg2-binary